Dtype-stable forward across Gemma 4 LLM + vision; vision scan body remat
All RMSNorm / Einsum / skip_scale operations in the Gemma 4 LLM and vision
modules now preserve the activation dtype (typically bf16) instead of
upcasting to f32 via implicit promotion against f32-stored params. Halves
activation HBM and matches the LLM's embed_dtype throughout. The vision
transformer scan body is also wrapped in nn.remat so the 16-layer scan
stops materialising the per-layer activation tape (~9 GB at B=256, 240x240).

Per-file changes:

- gemma/gm/nn/gemma4/_layers.py: Einsum casts w to x.dtype before matmul (params remain f32 in storage; matmul runs in activation precision); ClippedEinsum casts clip bounds to x.dtype so jnp.clip doesn't upcast a bf16 input; RMSNorm computes variance in f32 for stability and casts the result back to x.dtype.
- gemma/gm/nn/gemma4/_modules.py: skip_scale multiply uses .astype(outputs.dtype) (was upcasting the bf16 carry to f32 at every layer); per_layer_input_embedding_table lookup cast to x.dtype before scaling.
- gemma/gm/nn/gemma4/vision/_layers.py: VisionExit jnp.sqrt(d_model).astype(x.dtype) so the post-pool scaling stays in activation precision.
- gemma/gm/nn/gemma4/vision/_norms.py: RMSNorm dtype-stable (compute variance in f32, multiply by f32 scale, cast back to input dtype).
- gemma/gm/nn/gemma4/vision/_transformer.py: Wrap VisionBlock with nn.remat(prevent_cse=False, policy=nothing_saveable) inside the nn.scan so each iteration's intra-block intermediates are recomputed in backward instead of saved (~9 GB savings observed at B=256, 240x240).

+++ gemma/gm/nn/gemma4/_modules.py
@@ -181,7 +181,10 @@ class Embedder(nn.Module):
     )
     x = self.per_layer_model_projection('...td,dnp->...tnp', x)
     x = self.per_layer_projection_norm(x)
-    y = self.per_layer_input_embedding_table[(t,)]
+    # Embedding-table lookup returns the table's stored dtype (f32). Cast to
+    # the projection's activation dtype so the residual stream stays in bf16
+    # rather than being upcast by the f32 embedding.
+    y = self.per_layer_input_embedding_table[(t,)].astype(x.dtype)
     y *= jnp.sqrt(self.per_layer_input_dim).astype(y.dtype)
     return (x + y) * jax.lax.rsqrt(2.0).astype(x.dtype)
 
@@ -645,7 +648,9 @@ class Block(nn.Module):
       outputs += per_layer_inputs_mapped
 
     # 4. Scale
-    outputs = outputs * self.skip_scale
+    # Cast skip_scale (f32 param) to activation dtype so the per-layer carry
+    # stays in bf16 instead of being upcast to f32 by the multiplication.
+    outputs = outputs * self.skip_scale.astype(outputs.dtype)
 
     return cache, outputs, cls_attn_row
